requirements.txt
python-telegram-bot==20.7
vk-api==11.9.9
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
beautifulsoup4==4.12.3

